+++ src/CSS/profile.css
@@ -1,0 +1,3 @@
+.profile *{
+    background-color: #282c34;
+}
