+++ src/pages/HomePage.js
@@ -1,6 +1,6 @@
 import spinner from '../assets/spinner.svg';
 import logo from '../assets/roadmap.jpg';
-import '../App.css';
+import '../CSS/App.css';
 import Header from '../components/Header';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import HomePageAfterLogIn from './HomePageAfterLogIn';
