+++ src/pages/calendar.js
@@ -1,9 +1,9 @@
 import React from "react";
 import { useState } from 'react';
 import Calendar from 'react-calendar';
-//import 'react-calendar/dist/Calendar.css';
+// import 'react-calendar/dist/Calendar.css';
 import Header from "../components/Header";
-import calendar from "./calendar.css"
+// import calendar from "./calendar.css"
 import {Container, Row, Col} from 'react-bootstrap/';
