+++ src/pages/login.js
@@ -1,4 +1,4 @@
-import './LogIn.css';
+import '../CSS/LogIn.css';
 import Header from '../components/Header';
 import { Link } from 'react-router-dom';
 import { useEffect, useState } from 'react';
